the experimental version of batman. source files renamed to bmex.c.h to avoid confusion.
Although the core routing concept is similar to the main batman protocol
this thread incorporates a number of experimental features and a complete 
rewrite of the data structure maintained by each batman node to keep track 
of received originator messages and identified routes.

Also it offers additional information to a node (and its users) when trying to 
find optimal metrics (and hazards) for selecting the best next hop towards the final 
destination of a packet. 

The current version aims to supports (this is very experimental or work in progress):

- Multi interface support
- Routing table showing selected and alternative routes and characteristics to each destination
- Accessible characteristics for each potential route include:
  - Next-hop ip
  - Timeout for bidirectional neighbor status of next hop
  - Interface 
  - Number of total originator messages from destination received via this neighbor
  - Number of originator messages form destination received via this neighbor first
  - Average, minimm, and maximum TTL (Hops) of originator message
  - Minimum and maximum sequence number of received originator messages
  - Average, minimum, and maximum one-way trip time indicator which can be used to 
    identify the fastest of possibly multiple routes to a destination.
- Originator packet bundling to reduce MAC overhead when (re-)broadcasting own and other 
  nodes originator messages
- ...