Don't create different definitions of variables from header
Declaration of variables in header files must be marked extern or they will
create new definition in all files which include it. This could create
collisions on different architectures or by using different linkers.
A different effect is the misfeature of the program when different files access
these variable and think they would use the same variable in memory.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>