[PATCH] Introduce set/get functions for /proc files
get_integer_file can be used to get an integer from the beginning of a
file. It does the same as every other get_* function in linux/kernel.c
has done before, but introduced extra error checking for fscanf to
deal with the undefined behaviour of it in error cases in c89.

It fixes also the problems that in future architectures with
sizeof(int) > 4 an invalid write would happen when fscanf tries to write
it's results into the target or that on big endian architectures with
sizof(int) != 4 the result is complete different than the expected one
when the content of the file is not 0.

set_integer_file does the same for writing a specific integer to the
file and thus resolve the problem of invalid read on architectures with
sizof(int) > 4 and wrong file content on big endian systems with
sizof(int) != 4.

As side effect the probability to get attacked by a goto-raptor was
reduced by the factor 1.08.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>