[batman] Don't add size for netlink	header twice in netlink request
The parameter len of NLMSG_LENGTH is only the size of the payload of
the request. If we use this function to calculate the size of the buffer
we send to the kernel we would add (sizeof(nlmsghdr)+padding) and the
size of our own structure which also has a nlmsghdr included. We must
split the req structure into a header part and the payload to calculate
the correct size of our buffer and having the payload always start at
the correct address.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>