[batman] choose HNA gateway based on the TQ towards the gateway
This patch allows it to have multiple batman nodes in a single mesh that announce the same network. Each batman node will choose its best HNA gateway based on the TQ value. If the value changes and/or better gateway become available batman will switch to the new HNA gateway to ensure the best connection is used.