[batman-adv userspace] removing obsolete branch
The batman-adv userspace branch and its vis adv server are obsolete branches which have been superseded by the batman-adv kernelland branch (the layer 2 performance in user space is insufficient). These branches have been unmaintained for a while and increase the branch confusion for newcomers. If you wish to revive it please contact us.