[batctl] Prevent division by zero when no	packets were send
We will try to print a statistic even when we receive a sigterm or
sigint before we send any data or started to count the send data. This
means that the we try to calculate the data loss by dividing through
zero as it is the number of send packets. This is not possible and we
must check for that corner case.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>