batctl: Parse allowed settings for sysfs without whitespace before newline
The current settings parser assumes that a newline is at the end of each
line starting with "commands:". Luis de Bethencourt reported that this
is a bad coding style and should be avoided.

To parse the new settings lines we must also check for \n and not only
for single whitespaces. After each check it must be ensured that the old
character is placed again at the old place and the line seems to be
unmodified to the user.

For old versions with extra whitespace before a newline an extra check
must be applied to not allow commands with empty parameter like in
 batctl vis_mode ''

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>