batman-adv: layer2 unicast packet fragmentation
This patch implements a simple layer2 fragmentation to allow traffic exchange over
network interfaces with a MTU smaller than 1500 bytes. The fragmentation splits the
big packets into two parts and marks the frames accordingly. The receiving end
buffers the packets to reassemble the orignal packet before passing it to the higher
layers. This feature makes it necessary to modify the batman-adv encapsulation for
unicast packets by adding a sequence number, flags and the originator address.
This modifcation is part of a seperate packet type for fragemented packets to keep
the original overhead as low as possible. This patch enables the feature by default
to ensure the data traffic can travel through the network. But it also prints a
warning to notify the user about the performance implications.

Note: Fragmentation should be avoided at all costs since it has a dramatic impact on
the performance, especially when it comes wifi networks. Instead of a single packet,
2 packets have to be sent! Not only valuable airtime is wasted but also packetloss
decreases the throughput. A link with 50% packetloss and fragmentation enabled is
pretty much unusable.

Signed-off-by: Andreas Langer <an.langer at gmx.de>