batctl: traceroute - use all received packets to retrieve neighbor mac
Traceroute used the first packet to extract the mac address of a neighbor.
If the first packet was lost the mac address could not be retrieved
although the following 2 packets also contain it.
This patch uses the first received packet to extract the mac. If the
first packet is lost it uses the second or third.

Signed-off-by: Marek Lindner <lindner_marek@yahoo.de>