batman-adv: Don't call sysfs_del_hardif() in atomic context
sysfs_del_hardif invokes kobject_put, which might sleep. However, we
are not allowed to sleep during a call_rcu. There is also no need to
do the removal with an atomic call_rcu, as kobject_put only frees the
kobject when there is no more reference to it anyway.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>